split userlist into new comp
