+++ client/src/App.js
@@ -1,11 +1,10 @@
 import React from "react";
 import AdminPanel from "./components/Admin/AdminPanel/AdminPanel";
-import Calender from "./components/Shared/Calender";
 
 function App() {
   return (
     <div className="App">
-      <Calender />
+      <AdminPanel />
     </div>
   );
 }
